requirements.txt
# =============================================================================
# REQUIREMENTS.TXT - PHARMA DELIVERY
# =============================================================================
# Ce fichier liste toutes les dépendances Python requises pour le projet.
# Utilisez 'pip install -r requirements.txt' pour les installer.
# =============================================================================

# -----------------------------------------------------------------------------
# 1. CORE & ESSENTIELS
# -----------------------------------------------------------------------------
# Framework Flask et dépendances principales
Flask==2.3.3
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7

# -----------------------------------------------------------------------------
# 2. BASE DE DONNÉES
# -----------------------------------------------------------------------------
# ORM et connexion à la base de données
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9  # Pilote PostgreSQL
psycopg2-pool==1.1  # Gestion des connexions
SQLAlchemy-Utils==0.41.1  # Utilitaires supplémentaires

# Migrations de base de données
alembic==1.12.1
Flask-Migrate==4.0.5  # Intégration d'Alembic avec Flask

# -----------------------------------------------------------------------------
# 3. AUTHENTIFICATION ET AUTORISATION
# -----------------------------------------------------------------------------
# Gestion des sessions utilisateur
Flask-Login==0.6.2

# JSON Web Tokens (JWT) pour l'API
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0

# Hachage de mot de passe
passlib[bcrypt]==1.7.4

# -----------------------------------------------------------------------------
# 4. SÉCURITÉ
# -----------------------------------------------------------------------------
# CORS (Cross-Origin Resource Sharing)
Flask-Cors==4.0.0

# Sécurité HTTP (HSTS, CSP, etc.)
Flask-Talisman==1.1.0

# Limitation du taux de requêtes
Flask-Limiter==3.5.0

# Cryptographie
cryptography==41.0.7
pyOpenSSL==23.2.0
PyNaCl==1.5.0  # Pour le chiffrement

# -----------------------------------------------------------------------------
# 5. CONFIGURATION ET VARIABLES D'ENVIRONNEMENT
# -----------------------------------------------------------------------------
python-dotenv==1.0.0  # Chargement des variables d'environnement depuis .env
python-decouple==3.8  # Gestion de la configuration
dj-database-url==2.1.0  # Conversion des URL de base de données

# -----------------------------------------------------------------------------
# 6. FORMULAIRES ET VALIDATION
# -----------------------------------------------------------------------------
# Gestion des formulaires web
Flask-WTF==1.2.1
WTForms==3.0.1
email-validator==2.1.0  # Validation des emails

# Sérialisation/désérialisation
marshmallow==3.20.1
marshmallow-sqlalchemy==1.0.0
marshmallow-dataclass[enum,union]>=8.0.0

# -----------------------------------------------------------------------------
# 7. SERVEUR ET PERFORMANCE
# -----------------------------------------------------------------------------
# Serveur WSGI pour la production
gunicorn==21.2.0

# Support asynchrone (optionnel)
eventlet==0.33.3
gevent==23.7.0

# Cache
Flask-Caching==2.1.0
redis==5.0.1  # Backend pour le cache et les files d'attente

# Tâches en arrière-plan
celery==5.3.4
flower==2.0.1  # Interface d'administration pour Celery

# -----------------------------------------------------------------------------
# 8. EMAIL ET NOTIFICATIONS
# -----------------------------------------------------------------------------
# Envoi d'emails
Flask-Mail==0.9.1

# -----------------------------------------------------------------------------
# 9. MÉDIAS ET TÉLÉCHARGEMENTS
# -----------------------------------------------------------------------------
# Traitement des images
Pillow==10.1.0

# Détection du type de fichier
python-magic==0.4.27

# -----------------------------------------------------------------------------
# 10. TRAITEMENT DES DONNÉES
# -----------------------------------------------------------------------------
# Manipulation de données
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2  # Pour l'import/export Excel

# -----------------------------------------------------------------------------
# 11. DOCUMENTATION D'API
# -----------------------------------------------------------------------------
# Documentation interactive
flask-restx==1.1.0
apispec==6.4.0
apispec-webframeworks==3.0.0

# -----------------------------------------------------------------------------
# 12. SURVEILLANCE ET JOURNALISATION
# -----------------------------------------------------------------------------
# Surveillance des erreurs
sentry-sdk[flask]==1.34.0

# Journalisation structurée
python-json-logger==2.0.7
blinker==1.7.0  # Pour les signaux Flask

# -----------------------------------------------------------------------------
# 13. UTILITAIRES
# -----------------------------------------------------------------------------
# Requêtes HTTP
requests==2.31.0

# Manipulation des dates et heures
python-dateutil==2.8.2
pytz==2023.3.post1

# -----------------------------------------------------------------------------
# 14. DÉVELOPPEMENT ET TESTS (à ne pas installer en production)
# -----------------------------------------------------------------------------
# Test unitaire
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1  # Pour mocker les requêtes HTTP

# Formatage et vérification de code
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0

# Types pour mypy
types-requests==2.31.0.20240106
types-python-dateutil==2.8.19.14

# Internationalization
Babel==2.13.1
Flask-Babel==4.0.0
